albedo, illumination and normal renderer added
